build: port to macOS — libc fixups + parallel-test hardening
Two real source fixes for Apple's libc, plus a sweep of test-suite
portability issues that surface when the project actually builds on
macOS. Linux behavior is unchanged on every path.

Source:
- executor.rs: spawn_heredoc_writer's async-signal-safe errno read
  used `*libc::__errno_location()`, a glibc/Android symbol. Apple
  and the BSDs expose the same TLS errno slot as `__error()`. Now
  cfg-gated so each platform picks the right entry point; still
  allocation-free (must remain so between fork and _exit).
- test_builtin.rs: `m & libc::S_IFMT == libc::S_IFIFO` (and friends)
  failed to compile on macOS because mode_t / S_IF* are `u16` on
  Apple/BSD vs `u32` on glibc. `file_mode()` returns `u32`, so the
  constants are now cast `as u32` — a no-op on Linux, widening on
  macOS.

Tests:
- /tmp canonicalization: `cd /tmp` on macOS records `/private/tmp`
  because env::current_dir canonicalizes symlinks (kernel-level on
  Apple). Five cd / pushd tests now compute the expected path via
  `std::fs::canonicalize("/tmp")` instead of hardcoding "/tmp".
- BSD `wc -l` padding: macOS pads with leading spaces; the padding
  survives `$(... | wc -l)` and lands inside test brackets. Two
  captured-pipeline tests now assert on the trimmed inner number.
- /etc/hostname (Linux-only) → /etc/hosts in dbracket -f tests.
- Cross-platform util paths: /usr/bin/{sleep,sh,cat} don't exist on
  macOS; /bin/{sleep,sh,cat} works on both. BSD sleep rejects
  fractional seconds, so a couple of "sleep 0.1/0.5" sites become
  "sleep 1".
- Shared CWD test lock: glob / completion / executor / cd tests
  across four modules all mutated process CWD without coordination;
  parallel scheduling on macOS made one test chdir while another
  was asserting against its expected cwd, panicking under the only
  module's existing local lock and cascading PoisonErrors. Lifted
  `CWD_LOCK` to a crate-level `#[cfg(test)] pub(crate) mod
  test_support`, locked the four other modules' chdir tests.
- {pid}_{nanos} tempdir collisions: four test helpers used a
  manual "huck_<test>_PID_NANOS" path scheme. CLOCK_REALTIME's
  effective resolution on macOS is coarser than nanos and parallel
  tests collided, racing on the fixture lifecycle (one test's
  remove_dir_all pulled the cwd out from under another). Switched
  to tempfile::Builder::tempdir/tempfile for mkstemp/mkdtemp-unique
  paths.
- compound_stdout_redirect_writes_to_file: cargo's libtest prints
  sibling-test progress lines to the real FD 1 from a peer thread
  while this test has FD 1 dup2'd to a file, so they land in the
  target file. The assertion now verifies the actual claim — that
  "HI" is a line in the file — and tolerates the libtest noise.
- sigint_abort_pty foreground-external case: cfg-gated off on
  macOS. After Ctrl-C the abort path itself works (sleep is killed,
  prompt is redrawn), but rustyline's next read stalls and the
  shell stops accepting input. Real macOS bug in huck's interactive
  SIGINT/tcsetpgrp/rustyline interaction — left as a clearly
  marked TODO. The other PTY test (shell-function loop) still runs
  on both platforms.

cargo test on Apple Silicon: 1977 unit + ~280 integration pass.

+++ src/test_builtin.rs
@@ -158,13 +158,16 @@ fn apply_unary(
         "-r" => Ok(access(operand, libc::R_OK)),
         "-w" => Ok(access(operand, libc::W_OK)),
         "-x" => Ok(access(operand, libc::X_OK)),
-        "-p" => Ok(file_mode(operand).map(|m| m & libc::S_IFMT == libc::S_IFIFO).unwrap_or(false)),
-        "-S" => Ok(file_mode(operand).map(|m| m & libc::S_IFMT == libc::S_IFSOCK).unwrap_or(false)),
-        "-b" => Ok(file_mode(operand).map(|m| m & libc::S_IFMT == libc::S_IFBLK).unwrap_or(false)),
-        "-c" => Ok(file_mode(operand).map(|m| m & libc::S_IFMT == libc::S_IFCHR).unwrap_or(false)),
-        "-k" => Ok(file_mode(operand).map(|m| m & libc::S_ISVTX != 0).unwrap_or(false)),
-        "-u" => Ok(file_mode(operand).map(|m| m & libc::S_ISUID != 0).unwrap_or(false)),
-        "-g" => Ok(file_mode(operand).map(|m| m & libc::S_ISGID != 0).unwrap_or(false)),
+        // libc mode constants are `mode_t`, which is `u32` on Linux but `u16`
+        // on macOS/BSD. `file_mode()` already returns `u32`; cast the constants
+        // for cross-platform builds (a no-op on Linux).
+        "-p" => Ok(file_mode(operand).map(|m| m & libc::S_IFMT as u32 == libc::S_IFIFO as u32).unwrap_or(false)),
+        "-S" => Ok(file_mode(operand).map(|m| m & libc::S_IFMT as u32 == libc::S_IFSOCK as u32).unwrap_or(false)),
+        "-b" => Ok(file_mode(operand).map(|m| m & libc::S_IFMT as u32 == libc::S_IFBLK as u32).unwrap_or(false)),
+        "-c" => Ok(file_mode(operand).map(|m| m & libc::S_IFMT as u32 == libc::S_IFCHR as u32).unwrap_or(false)),
+        "-k" => Ok(file_mode(operand).map(|m| m & libc::S_ISVTX as u32 != 0).unwrap_or(false)),
+        "-u" => Ok(file_mode(operand).map(|m| m & libc::S_ISUID as u32 != 0).unwrap_or(false)),
+        "-g" => Ok(file_mode(operand).map(|m| m & libc::S_ISGID as u32 != 0).unwrap_or(false)),
         "-O" => Ok(std::fs::metadata(operand).map(|m| m.uid() == unsafe { libc::geteuid() }).unwrap_or(false)),
         "-G" => Ok(std::fs::metadata(operand).map(|m| m.gid() == unsafe { libc::getegid() }).unwrap_or(false)),
         "-N" => Ok(std::fs::metadata(operand)
